[FEAT]회원탈퇴 API 구현 완료

+++ app/src/models/User.js
@@ -48,9 +48,6 @@ class User{
         const client = this.body;
         let userInfo=await UserStorage.getUserInfo(client.user_email);
        
-        console.log("입력한 psword:",client.psword);
-        console.log("기존 psword:",userInfo.psword);
-        // if(client.psword===userInfo.psword){
         if(await bcrypt.compare(client.psword,userInfo.psword)){
             const response = await UserStorage.updatePsword(client);
             return response;
@@ -67,8 +64,18 @@ class User{
     //회원 탈퇴
     async withdrawalUser(){
         const client = this.body;
-        const response = await UserStorage.deleteUser(client);
-        return response;
+        let userInfo=await UserStorage.getUserInfo(client.user_email);
+       
+        if(await bcrypt.compare(client.psword,userInfo.psword)){
+            const response = await UserStorage.deleteUser(client);
+            return response;
+        }else{
+            return {
+                result:false,
+                status: 400,
+                err: `비밀번호가 틀렸습니다.`
+            }
+        }
     }
